+++ cogs/github/base/org.py
@@ -1,7 +1,7 @@
 import discord
 import datetime
 from typing import Optional, Union
-from cfg import bot_config
+from core import bot_config
 from discord.ext import commands
 
 Git = bot_config.Git
