+++ cogs/github/base/user.py
@@ -2,7 +2,7 @@ import discord
 import datetime
 from discord.ext import commands
 from typing import Union, Optional
-from cfg import bot_config
+from core import bot_config
 
 Git = bot_config.Git
